% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perseus_text.R
\name{get_perseus_text}
\alias{get_perseus_text}
\title{Get a primary text by urn.}
\usage{
get_perseus_text(urn, language = "eng", text = NULL)
}
\arguments{
\item{urn}{resource number obtained from perseus_catalog}

\item{language}{which language to be returned. "grc" for Greek, "lat" for Latin, and "eng" for English.}

\item{text}{a precise text citation, e.g. '1.1-1.10'. If left NULL, the whole work is returned.}
}
\value{
a six column tbl_df
}
\description{
Get a primary text by urn.
}
\examples{
\dontrun{
get_perseus_text(urn = "urn:cts:greekLit:tlg0031.tlg013", language = "grc", text = "1.1-1.10")
get_perseus_text(urn = "urn:cts:greekLit:tlg0031.tlg002")
get_perseus_text(urn = "urn:cts:latinLit:stoa0104p.stoa006", language = "lat")
}
}
