% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_excerpt.R
\name{parse_excerpt}
\alias{parse_excerpt}
\title{Parse a Greek excerpt}
\usage{
parse_excerpt(urn, excerpt)
}
\arguments{
\item{urn}{a valid urn from the perseus_catalog object.}

\item{excerpt}{a valid excerpt, e.g. 5.1-5.5}
}
\value{
a data frame
}
\description{
This function parses a Greek excerpt from the Perseus Digital Library. Parsing includes
part of speech, gender, case, mood, voice, tense, person, number, and degree.
}
\examples{
parse_excerpt("urn:cts:greekLit:tlg0031.tlg002.perseus-grc2", "5.1-5.4")
}
