% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perseus_parallel.R
\name{perseus_parallel}
\alias{perseus_parallel}
\title{Render a text parallel with ggplot2}
\usage{
perseus_parallel(perseus_df, words_per_row = 6)
}
\arguments{
\item{perseus_df}{a data frame obtained from \code{get_perseus_text}. Can contain multiple texts.}

\item{words_per_row}{adjusts the words displayed per "row".}
}
\value{
a ggplot object
}
\description{
Render a text parallel with ggplot2
}
\examples{
\dontrun{
tibble::tibble(label = c("Colossians", rep("1 Thessalonians", 2), "Romans"),
               excerpt = c("1.4", "1.3", "5.8", "8.35-8.39")) \%>\%
 dplyr::left_join(perseus_catalog) \%>\%
 dplyr::filter(language == "grc") \%>\%
 dplyr::select(urn, excerpt) \%>\%
 as.list() \%>\%
 purrr::pmap_df(get_perseus_text) \%>\%
 perseus_parallel()
}
}
