% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perseus_catalog.R
\name{get_perseus_catalog}
\alias{get_perseus_catalog}
\title{Get the Perseus Catalog}
\usage{
get_perseus_catalog()
}
\value{
catalog data frame
}
\description{
Returns the Perseus catalog, including urn lookups. Note: this function
is no longer exported. The catalog is now included in /data and is lazily loaded.
You can do a manual search of the catalog here: http://cts.perseids.org/
}
\examples{
\dontrun{
get_perseus_catalog()
}
}
