% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perseus_text.R
\name{get_perseus_text}
\alias{get_perseus_text}
\title{Get a primary text by URN.}
\usage{
get_perseus_text(text_urn)
}
\arguments{
\item{text_urn}{Valid uniform resource number (URN) obtained from perseus_catalog.}
}
\value{
a seven column tbl_df
}
\description{
Get a primary text by URN.
}
\examples{
\dontrun{
get_perseus_text("urn:cts:greekLit:tlg0013.tlg028.perseus-eng2")
get_perseus_text("urn:cts:greekLit:tlg0013.tlg028.perseus-grc2")
get_perseus_text("urn:cts:latinLit:stoa0215b.stoa003.opp-lat1")
}
}
