% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perseus_text.R
\name{get_perseus_text}
\alias{get_perseus_text}
\title{Get a primary text by urn}
\usage{
get_perseus_text(urn, text)
}
\arguments{
\item{urn}{resource identifier obtained from get_perseus_catalog()}

\item{text}{a precise text citation, e.g. '1.1-1.10'}
}
\value{
character vector of primary text
}
\description{
Get a primary text by urn
}
\examples{
get_perseus_text("urn:cts:greekLit:tlg0031.tlg013", "1.1-1.10")
}

