% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perseus_text.R
\name{get_perseus_text}
\alias{get_perseus_text}
\title{Get a primary text by URN.}
\usage{
get_perseus_text(text_urn)
}
\arguments{
\item{text_urn}{Valid uniform resource number (URN) obtained from perseus_catalog.}
}
\value{
A seven column tbl_df with one row for each "section" (splits vary from text--could be line, chapter, etc.).
Columns:
\describe{
  \item{text}{character vector of text}
  \item{urn}{Uniform Resource Number}
  \item{group_name}{Could refer to author (e.g. "Aristotle") or corpus (e.g. "New Testament")}
  \item{label}{Text label, e.g. "Phaedrus"}
  \item{description}{Text description}
  \item{language}{Text language, e.g. "grc" = Greek, "lat" = Latin, "eng" = English}
}
}
\description{
Get a primary text by URN.
}
\examples{
\dontrun{
get_perseus_text("urn:cts:greekLit:tlg0013.tlg028.perseus-eng2")
get_perseus_text("urn:cts:greekLit:tlg0013.tlg028.perseus-grc2")
get_perseus_text("urn:cts:latinLit:stoa0215b.stoa003.opp-lat1")
}
}
